+++ src/DroidSolutions.Oss.JobService.EFCore/IJobContext.cs
@@ -7,14 +7,10 @@ namespace DroidSolutions.Oss.JobService.EFCore;
 /// <summary>
 /// A data context that integrates a generic job table.
 /// </summary>
-/// <typeparam name="TParams">The type of the job parameters.</typeparam>
-/// <typeparam name="TResult">The type of the job result.</typeparam>
-public interface IJobContext<TParams, TResult>
-  where TParams : class?
-  where TResult : class?
+public interface IJobContext
 {
   /// <summary>
   /// Gets the job table.
   /// </summary>
-  public DbSet<Job<TParams, TResult>> Jobs { get; }
+  public DbSet<JobBase> Jobs { get; }
 }