+++ src/NodeJS/Index.ts
@@ -2,6 +2,7 @@ export type { IJob } from "./Generated/IJob";
 export type { IJobRepository } from "./Generated/IJobRepository";
 export { JobState } from "./Generated/JobState";
 export type { IJobWorkerBase } from "./Generated/Worker/IJobWorkerBase";
+export { IJobWorkerMetrics } from "./Generated/Worker/IJobWorkerMetrics";
 export type { IJobWorkerSettings } from "./Generated/Worker/Settings/IJobWorkerSettings";
 export { EmtpyLogger } from "./Helper/LoggerFactory";
 export type { LoggerFactory, SimpleLogger } from "./Helper/LoggerFactory";