+++ test/DroidSolutions.Oss.JobService.EFCore.Test/Fixture/InMemoryJobRepository.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Diagnostics;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
@@ -20,7 +19,7 @@ public class InMemoryJobRepository : JobRepositoryBase<TestContext, TestParamete
   }
 
   public override async Task AddProgressAsync(
-    IJob<TestParameter, TestResult> job,
+    IJobBase job,
     int items = 1,
     bool failed = false,
     CancellationToken cancellationToken = default)
@@ -43,10 +42,21 @@ public class InMemoryJobRepository : JobRepositoryBase<TestContext, TestParamete
     string type,
     DateTime? dueDate,
     TestParameter? parameters,
+    bool includeStarted = false,
     CancellationToken cancellationToken = default)
   {
     IQueryable<Job<TestParameter, TestResult>> query = Context.Jobs
-      .Where(x => x.Type == type && x.State == JobState.Requested);
+      .Where(x => x.Type == type)
+      .Cast<Job<TestParameter, TestResult>>();
+
+    if (includeStarted)
+    {
+      query = query.Where(x => x.State == JobState.Started || x.State == JobState.Requested);
+    }
+    else
+    {
+      query = query.Where(x => x.State == JobState.Requested);
+    }
 
     if (dueDate.HasValue)
     {
@@ -71,6 +81,7 @@ public class InMemoryJobRepository : JobRepositoryBase<TestContext, TestParamete
   {
     Job<TestParameter, TestResult>? job = await Context.Jobs
       .OrderBy(x => x.DueDate)
+      .Cast<Job<TestParameter, TestResult>>()
       .FirstOrDefaultAsync(
         x => x.State == JobState.Requested && x.Type == type && x.DueDate <= DateTime.UtcNow,
         cancellationToken);